Home Management Component

+++ backend/src/main/java/com/tfg/backend/controller/ShopRestController.java
@@ -7,17 +7,19 @@ import com.tfg.backend.utils.GlobalDefaults;
 import com.tfg.backend.utils.PageFormatter;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.io.IOException;
+import java.net.URI;
 import java.util.*;
 
 @RestController
@@ -81,6 +83,18 @@ public class ShopRestController {
     }
 
 
+    @Operation(summary = "(Driver) Get shop information by assigned truck ID")
+    @GetMapping("/truck/{id}")
+    public ResponseEntity<ShopDTO> getShopByAssignedTruckId(@PathVariable Long id) {
+        Truck truck = truckService.findTruckHelper(id);
+        Shop assignedShop = truck.getAssignedShop();
+        if (assignedShop == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "This truck is not assigned to any shop.");
+        }
+        return ResponseEntity.ok(new ShopDTO(assignedShop));
+    }
+
+
     @Operation(summary = "(All) Get shop stock page by ID")
     @GetMapping("/stock/{id}")
     public ResponseEntity<PageResponse<ShopStockDTO>> getShopStocks(@PathVariable Long id, Pageable pageable) {
@@ -96,9 +110,16 @@ public class ShopRestController {
         Address address = new Address(dto.getAlias(), dto.getStreet(), dto.getNumber(), dto.getFloor(), dto.getPostalCode(), dto.getCity(), dto.getCountry());
         address.setLatitude(shopDTO.getAddress().getLatitude());
         address.setLongitude(shopDTO.getAddress().getLongitude());
-        Shop shop = new Shop(shopDTO.getName(), address);
+        Shop shop = new Shop(shopDTO.getName(), address, shopDTO.getAssignedBudget());
         Shop savedShop = shopService.save(shop);
-        return ResponseEntity.accepted().body(new ShopDTO(savedShop));
+
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(savedShop.getId())
+                .toUri();
+
+        return ResponseEntity.created(location).body(new ShopDTO(savedShop));
     }
 
     @Operation(summary = "(Admin) Update shop by ID")
@@ -112,27 +133,48 @@ public class ShopRestController {
         address.setLatitude(dto.getLatitude());
         address.setLongitude(dto.getLongitude());
         shop.setAddress(address);
+        shop.setAssignedBudget(shopDTO.getAssignedBudget());
         Shop updatedShop = shopService.save(shop);
 
         return ResponseEntity.accepted().body(new ShopDTO(updatedShop));
     }
 
 
+    @Transactional
     @Operation(summary = "(Admin) Delete shop by ID")
     @DeleteMapping("/{id}")
     public ResponseEntity<ShopDTO> deleteShop(@PathVariable Long id) {
         Shop shop = shopService.findShopHelper(id);
 
-        //Unlink trucks
-        List<Truck> assignedTrucks = shop.getAssignedTrucks();
+        // Unlink trucks
+        List<Truck> assignedTrucks = new ArrayList<>(shop.getAssignedTrucks());
         for (Truck truck : assignedTrucks) {
             truck.setAssignedShop(null);
         }
-        truckService.saveAll(assignedTrucks);
+        shop.getAssignedTrucks().clear();
+
+        // Unlink orders
+        List<Order> assignedOrders = new ArrayList<>(shop.getAssignedOrders());
+        for (Order order : assignedOrders) {
+            order.setAssignedShop(null);
+
+            if(order.getCurrentStatus() == OrderStatus.ORDER_MADE || order.getCurrentStatus() == OrderStatus.SENT){
+                order.changeOrderStatus(OrderStatus.CANCELLED, "La tienda a la que estaba asignado el pedido ha sido eliminada.");
+            }
+        }
+        shop.getAssignedOrders().clear();
+
+        // Unlink clients (that have this shop as selected)
+        List<User> customers = new ArrayList<>(shop.getCustomers());
+        for (User customer : customers) {
+            customer.setSelectedShop(null);
+        }
+        shop.getCustomers().clear();
 
         shopService.delete(shop);
-        //Delete shop image (if it is not the default photo)
-        if (!shop.getImage().equals(GlobalDefaults.SHOP_IMAGE)) {
+
+        // Delete shop image if it is not the default image
+        if (shop.getImage() != null && !shop.getImage().equals(GlobalDefaults.SHOP_IMAGE)) {
             storageService.deleteFile(shop.getImage().getS3Key());
         }
 
@@ -141,7 +183,7 @@ public class ShopRestController {
 
 
     @Operation(summary = "(Manager) Toggle stock local activation by ID")
-    @PostMapping("/active/{id}")
+    @PutMapping("/active/{id}")
     public ResponseEntity<ShopStockDTO> toggleLocalActivation(@PathVariable Long id, @RequestParam boolean state) {
         ShopStock stock = shopStockService.findShopStockHelper(id);
         stock.setActive(state);
@@ -151,7 +193,7 @@ public class ShopRestController {
 
 
     @Operation(summary = "(Manager) Toggle all stocks local activation")
-    @PostMapping("/{shopId}/active/")
+    @PutMapping("/{shopId}/active/")
     public ResponseEntity<Boolean> toggleAllLocalActivations(@PathVariable Long shopId, @RequestParam boolean state) {
         List<ShopStock> stocks = this.shopStockService.findAllByShopId(shopId);
         for (ShopStock s : stocks) {
@@ -163,20 +205,30 @@ public class ShopRestController {
 
 
     @Operation(summary = "(Manager) Add n units to a shop's product stock (if exists)")
-    @PostMapping("/restock/{stockId}")
+    @PutMapping("/restock/{stockId}")
     public ResponseEntity<ShopStockDTO> restockProduct(@PathVariable Long stockId, @RequestParam int units) {
         ShopStock targetStock = shopStockService.findShopStockHelper(stockId);
-        if (units > 0){
-            targetStock.setUnits(targetStock.getUnits() + units);
-            ShopStock savedStock = shopStockService.save(targetStock);
-            return ResponseEntity.ok(new ShopStockDTO(savedStock));
+
+        if (units <= 0){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Restock units must be positive, " + units + " is not a valid number.");
         }
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Restock units must be positive, " + units + " is not a valid number.");
+
+        Shop restockingShop = targetStock.getShop();
+        double supplyCost = units * targetStock.getProduct().getSupplyPrice();
+        if (supplyCost > restockingShop.getAssignedBudget()){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "There is not enough budget in this shop to complete this operation.");
+        }
+
+        targetStock.setUnits(targetStock.getUnits() + units);
+        restockingShop.setAssignedBudget(restockingShop.getAssignedBudget() - supplyCost);
+        shopService.save(restockingShop);
+        ShopStock savedStock = shopStockService.save(targetStock);
+        return ResponseEntity.ok(new ShopStockDTO(savedStock));
     }
 
 
     @Operation(summary = "(Manager) Set stock assignment to a shop")
-    @PostMapping("/{shopId}/assign/stock/{stockId}")
+    @PutMapping("/{shopId}/assign/stock/{stockId}")
     public ResponseEntity<ShopStockDTO> setAssignedStock(@PathVariable Long shopId, @PathVariable Long stockId, @RequestParam boolean state) {
         Shop shop = shopService.findShopHelper(shopId);
         ShopStock targetStock;
@@ -196,7 +248,7 @@ public class ShopRestController {
 
 
     @Operation(summary = "(Manager) Set truck assignment to a shop")
-    @PostMapping("/{shopId}/assign/truck/{truckId}")
+    @PutMapping("/{shopId}/assign/truck/{truckId}")
     public ResponseEntity<TruckDTO> setAssignedTruck(@PathVariable Long shopId, @PathVariable Long truckId, @RequestParam boolean state) {
         Shop shop = shopService.findShopHelper(shopId);
         Truck truck = truckService.findTruckHelper(truckId);
@@ -214,7 +266,7 @@ public class ShopRestController {
 
 
     @Operation(summary = "(Admin) Set manager assignment to a shop")
-    @PostMapping("/{shopId}/assign/manager/{userId}")
+    @PutMapping("/{shopId}/assign/manager/{userId}")
     public ResponseEntity<ShopDTO> setAssignedManager(@PathVariable Long shopId, @PathVariable Long userId, @RequestParam boolean state) {
         Shop shop = shopService.findShopHelper(shopId);
 
@@ -235,7 +287,7 @@ public class ShopRestController {
 
 
     @Operation(summary = "(Admin) Update remote shop image")
-    @PostMapping("/image/{id}")
+    @PutMapping("/image/{id}")
     public ResponseEntity<ShopDTO> uploadShopImage(@PathVariable Long id, @RequestParam(value = "image", required = false) MultipartFile image) throws IOException {
         Shop shop = shopService.findShopHelper(id);
